Enhancements ✨
## 1. Optimisations et nettoyage
J’ai fait un bon coup de balai dans le code ! J’ai retiré tout ce qui était déprécié, comme certains événements ou bibliothèques. Par exemple, j’ai remplacé les vieux événements Checked/Unchecked par IsCheckedChanged, qui est plus moderne et propre. J’ai aussi supprimé des fichiers et des classes en double pour alléger un peu tout ça.

## 2. Mises à jour des dépendances
J’ai ajouté une nouvelle bibliothèque (Avalonia.Controls.ItemsRepeater) pour améliorer l’interface utilisateur. Ça va me permettre de proposer des trucs plus sympas et plus efficaces côté affichage.

## 3. Renommage pour plus de clarté
J’ai revu certains noms de classes et de méthodes pour qu’ils soient plus explicites. Par exemple, la classe Settings s’appelle maintenant AppSettings, histoire qu’on sache tout de suite de quoi il s’agit. Pareil pour des méthodes comme SelectJavaPathAsync, qui est désormais plus descriptive.

## 4. Améliorations de l’interface utilisateur
J’ai bossé sur l’UI pour la rendre plus moderne et agréable :
- Ajout de styles et d’animations pour les boutons, onglets, etc.
- Nouvelle organisation des paramètres : gestion de la mémoire avec un slider, choix du thème avec des aperçus visuels, et même la possibilité d’ajouter des arguments JVM personnalisés.
- J’ai aussi supprimé les styles inutilisés pour garder quelque chose de propre.

## 5. Gestion des paramètres
Les paramètres sont maintenant mieux organisés et sauvegardés dans un fichier JSON (settings.json). J’ai aussi mis en place une nouvelle détection automatique de Java qui va chercher dans les variables d’environnement, le PATH, et d’autres endroits courants.

## 6. Corrections de bugs
J’ai corrigé pas mal de petits soucis, comme des exceptions qui pouvaient faire planter l’appli (par exemple, lors de la détection de Java ou l’ouverture d’URL). J’ai aussi renforcé les vérifications pour éviter les mauvaises surprises.

## 7. Suppression de code inutile
Enfin, j’ai supprimé tout ce qui ne servait plus, comme la classe SimpleSettingsWindow, que j’ai fusionnée avec SettingsWindow.
