+++ MCLauncher.csproj
@@ -19,6 +19,7 @@
         </PackageReference>
         <PackageReference Include="CmlLib.Core" Version="4.0.5"/>
         <PackageReference Include="Avalonia.Xaml.Behaviors" Version="11.0.6"/>
+        <PackageReference Include="Avalonia.Controls.ItemsRepeater" Version="11.0.5"/>
     </ItemGroup>
 
     <ItemGroup>
